backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.3.0